backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pdfplumber==0.11.7
python-docx==1.2.0
spacy==3.8.7
requests==2.32.5
sentence-transformers==5.1.0
scikit-learn==1.7.1
SQLAlchemy==2.0.43
PyMySQL==1.1.1
